Moving existing tests into NamedDayTests folder
