+++ src/TodoTests/NamedDayTests/EasterTests.cs
@@ -2,7 +2,7 @@ using System;
 using NUnit.Framework;
 using Todo.Dates.Naming;
 
-namespace TodoTests;
+namespace TodoTests.NamedDayTests;
 
 public class EasterTests
 {
